+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- merge reflect-config.json utility
+
 ## [1.3.4] - 2024-04-01
 
 ### Fixed